fix: fix bug related to navbar and titles
Problem : When clicking on on link in the navbar, we were send to the corresponding section of the link, but the section title was hidden beneath the navbar.

Fix : I fixed the problem by adding an empty div just before the title, and giving it the id="section" corresponding to the section, so that this div could be hidden by the navbar, but that the title can still be visible.